pyproject.toml
[tool.poetry]
name = "langchainlab"
version = "0.1.0"
description = ""
authors = ["Firaol Bogale <firaolbogala98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
langchain = "^0.3.25"
langchain-google-genai = "^2.1.4"
python-dotenv = "^1.1.0"
langchain-community = "^0.3.24"
pypdf = "^5.5.0"
langchain-google-vertexai = "^2.0.23"
langchain-chroma = "^0.2.3"
requests = "^2.32.3"
langgraph = ">0.2.27"
bs4 = "^0.0.2"

[project]
requires-python = ">=3.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
